requirements.txt
Django==1.11.6
django-filter==1.1.0
olefile==0.44
Pillow==4.3.0
pkg-resources==0.0.0
psycopg2==2.7.3.2
pytz==2017.2
